tests: began implementing test for pkg_resources module content listing
The test script tests for behavior of resource_exists(), resource_isdir(),
and resource_listdir() functions from pkg_resources package (which
in turn call the methods with same name in the provider class).

The idea is to run the test script twice; first as native python
script, and then as a frozen program. The test module is present
as either plain module directory, an unzipped egg, or zipped egg.
This way, we test behavior of the original provider (DefaultProvider
or ZipProvider) and the provider used within the frozen application
(which we will need to implement to replace the currently used
NullProvider).