hooks: pkg_resources: begin implementing custom provider
Implement PyiFrozenProvider, a simple subclass of NullProvider
from pkg_resources that provides _has(), _isdir(), and _listdir()
methods. Their implementations are currently equivalent to the
ones in DefaultProvider and are filesystem-only.

This should be enough to make pkg_resources.resource_listdir()
work with the non-egg modules. However, it is not sufficient
for the conformance test, because it does not implement support
for embedded resources.