hooks: pkg_resources: do not resolve symlinks in input path
Avoid using Path.resolve() to normalize the input path, because
it resolves symlinks. The top-level package path does not have
symlinks resolved, so verification that the given path is located
within the package may fail if there's a symlink involved.

For now, use os.path.abspath() to normalize the input path.

Fixes failing onefile tests on macOS, where temporary directory is
placed in /var, which is actually a symlink to /private/var.