hooks: PyQt5/PySide2: relocate the Qt*.framework bundles
This is required for QtWebEngine sandboxing, which otherwise
prevents loading of resources in onefile builds. But for
consistency sake, relocate all Qt*.framework bundles (which
mirrors the location in the unfrozen application, as well).

The QtWebEngine/chromium sandboxing seems to allow loading of
resources from either application path or Qt library path.
In onedir builds, the application path corresponds to the _MEIPASS
dir, and the Qt library path corresponds to the
_MEIPASS/PySide2/Qt/lib or _MEIPASS/PyQt5/Qt/lib. Therefore,
in onedir builds, it does not matter if framework bundles (and
their resources) are placed directly in the _MEIPASS dir.

In onefile builds, however, the application path does not
correspond to the _MEIPASS dir anymore. Therefore, the only valid
location for the resources becomes the Qt library path, and hence
we need to relocate the framework bundles in there. Otherwise,
the main process loads the resources correctly, while the helper
QtWebEngineProcess processes will fail to mmap the icudtl.dat
file.