bootloader: archive: fix cookie search on macOS in unsigned binary
When signature is not found, we need to return (literal) -1, instead
of -1 stored in uint32_t.

Before any size-related modifications, the function returned signed
int, so the value ended up "correctly" cast back into -1. Then when
we changed return type to unsigned int, and all was well, because
0xFFFFFFFF was returned, and was later on correctly interpreted
by the caller (due to comparison against signed -1, which again
ended up "correct").

Now that the function returns int64_t, however, the value is
interpreted as actual offset 0x00000000FFFFFFFF, and therefore
we end up looking for the cookie in the wrong place. So we need
to actually return -1.