bootloader: remove MSC-specific use of _snprintf() and _vsnprintf()
Remove the Visual Studio specific mappings of snprintf() and
vsnprintf() to _snprintf() and _vsnprintf(). The support for actual
snprintf() and vsnprintf() is available in VS2015 and later.

Fixes the failing bootloader tests involving paths with lengths
exceeding PATH_MAX.

The problem is that behavior of _snprintf() is not compatible with
that of snprintf(); on overflow, it does not null-terminate the
string, and it returns -1 instead of the number of characters that
would have been written.