bootloader: use snprintf() to format full message in mbfatal_*error()
Replace the per-partes formatting of message-to-be-displayed in
mbfatal_winerror() and mbfatal_perror() with a single snprintf()
call.

This fixes a potential bug - the size check before the
the strncpy() call that appends the two-character ": "
to message ensures only that the message buffer can fit
one more character instead of two. So in a corner case,
this could result in an out-of-boundary write by one byte.

Also resolves a compiler warning when building with MinGW,
which was due to a strncpy() call that (likely intentionally)
did not copy the the terminating null from a literal ": " string.